initial Search and View filters
