+++ src/Nox.Generator.Tasks/Ui/Data/ApiUiServiceGenerator.cs
@@ -1,0 +1,30 @@
+﻿using Nox.Generator.Common;
+using Nox.Solution;
+
+namespace Nox.Generator.Tasks.Ui.Components;
+
+internal class ApiUiServiceGenerator : INoxFileGenerator
+{
+    public NoxGeneratorKind GeneratorKind => NoxGeneratorKind.Ui;
+
+    public void Generate(
+      NoxCodeGenConventions codeGeneratorState,
+      GeneratorConfig config,
+      System.Action<string> log,
+      string absoluteOutputPath
+      )
+    {
+
+        if (codeGeneratorState.Solution.Domain is null)
+        {
+            return;
+        }
+
+        var templateName = @"Tasks.Ui.Data.ApiUiService";
+
+        new TaskTemplateFileBuilder(codeGeneratorState, absoluteOutputPath)
+            .WithClassName($"ApiUiService")
+            .WithFileNamePrefix($"Ui.Data")
+            .GenerateSourceCodeFromResource(templateName);
+    }
+}
