+++ src/Nox.Generator.Tasks/Ui/Data/Paging.scriban-cs
@@ -1,4 +1,4 @@
-﻿namespace {{codeGenConventions.UiNameSpace}}.Configuration;
+﻿namespace {{codeGenConventions.UiNameSpace}}.Data;
 
 /// <summary>
 /// Setting class to define Pagination filter options used on returned result entities from target Api in the Ui
