+++ src/Nox.Generator.Tasks/Ui/Data/SortOrder.scriban-cs
@@ -1,6 +1,6 @@
 ﻿using {{codeGenConventions.UiNameSpace}}.Enum;
 
-namespace {{codeGenConventions.UiNameSpace}}.Configuration;
+namespace {{codeGenConventions.UiNameSpace}}.Data;
 
 /// <summary>
 /// Setting class to define column order options used on returned result entities from target Api in the Ui
