+++ src/Nox.Generator.Tasks/Ui/DataGrid/EntityDataGrid.scriban-cs
@@ -1,43 +1,122 @@
 ﻿using Microsoft.AspNetCore.Components;
-
 using {{codeGenConventions.UiNameSpace}}.Models;
+using {{codeGenConventions.UiNameSpace}}.Services;
 
 namespace {{codeGenConventions.UiNameSpace}}.DataGrid;
 
 public partial class {{className}} : ComponentBase
 {
-    private List<{{entity.Name}}Model> {{entity.PluralName}} = new(); 
+    private List<{{entity.Name}}Model> {{entity.PluralName}}Data = new(); 
+
+    public EntityData<{{entity.Name}}Model>? {{entity.Name}}EntityData { get; set; }
+
+    public IEnumerable<{{entity.Name}}Model>? PagedData = null;
+
+    public MudTable<{{entity.Name}}Model>? {{entity.Name}}DataGridTable;
+
     private bool IsLoading = true;
+
+    [Parameter]
+    public ApiUiService CurrentApiUiService { get; set; } = new();
 
     [Parameter]
     public EventCallback<{{entity.Name}}Model?> OnSelectionChanged { get; set; }
 
     [Parameter]
-    public EventCallback<{{entity.Name}}Model?> OnDeleteChanged { get; set; }
+    public EventCallback<{{entity.Name}}Model?> OnDeleteChanged { get; set; }    
+
+    private string PreviousAPiQuery = string.Empty;
 
     protected override async Task OnInitializedAsync()
     {
-        await LoadData();
+        IsLoading = true;
+
+        CurrentApiUiService.ResetAllSearchFilterList();
+        CurrentApiUiService.ResetOrderList();  
+        CurrentApiUiService.ResetShowInSearchList();
+
+
+
+        IsLoading = false;
     }
 
-    public async Task LoadData()
+    #region Data
+
+    public async Task<TableData<T>> ServerReload(TableState State)
     {
-        try
+        if (IsPagingPopulated)
         {
-            IsLoading = true;
-            {{entity.PluralName}} = await {{entity.PluralName}}Service.GetAllAsync();
+            CurrentApiUiService!.Paging!.SetCurrentPage(State.Page);
+            CurrentApiUiService!.Paging!.SetCurrentPageSize(State.PageSize);
         }
-        finally
+
+        await GetApiEntityData();
+
+        return new TableData<T>() { TotalItems = Convert.ToInt32(ApiEntityData?.EntityTotal), Items = PagedData };
+    }
+
+    public bool IsPagingPopulated
+    {
+        get
         {
-            IsLoading = false;
+            return CurrentApiUiService.Paging != null;
         }
     }
 
-    public async Task RefreshDataGrid()
+    private async Task GetApiEntityData()
     {
-        await LoadData();
-        StateHasChanged();
+        string CurrentApiQuery = CurrentApiUiService!.ApiGetQuery;
+
+        if (ApiEntityData == null
+            || String.IsNullOrEmpty(PreviousAPiQuery)
+            || !String.Equals(PreviousAPiQuery, CurrentApiQuery, StringComparison.OrdinalIgnoreCase))
+        {
+            ApiEntityData = await EntityDataService<T>.GetAsyncEntityData(CurrentApiUiService);
+
+            PreviousAPiQuery = CurrentApiQuery;
+
+            if (IsPagingPopulated)
+            {
+                CurrentApiUiService!.Paging!.SetEntityTotal(ApiEntityData?.EntityTotal);
+            }
+
+            PagedData = ApiEntityData?.EntityList;
+        }
     }
+
+    #endregion
+
+    #region Search
+
+    public bool IsSearchFilterPopulated
+    {
+        get
+        {
+            return CurrentApiUiService.SearchFilterList != null
+                && CurrentApiUiService.SearchFilterList.Any();
+        }
+    }
+
+    public async Task Search()
+    {
+        IsLoading = true;
+
+        if (IsPagingPopulated)
+        {
+            CurrentApiUiService!.Paging!.ResetPaging();
+        }
+
+        if ({{entity.Name}}DataGridTable != null)
+        {
+            await {{entity.Name}}DataGridTable!.ReloadServerData();
+        }
+
+        IsLoading = false;
+    }
+
+    #endregion
+
+    #region Edit
 
     public async Task SelectedOnClick({{entity.Name}}Model? currentSelection)
     {
@@ -47,6 +126,82 @@ public partial class {{className}} : ComponentBase
         }        
     }
 
+    #endregion
+
+    #region Order
+
+    public bool IsOrderPopulated
+    {
+        get
+        {
+            return CurrentApiUiService.OrderList != null
+                    && CurrentApiUiService.OrderList.Any();
+        }
+    }
+
+    public async Task UpdateOrder(string OrderTypeValue, string PropertyName)
+    {
+        IsLoading = true;
+
+        if (!string.IsNullOrWhiteSpace(PropertyName)
+            && IsOrderPopulated)
+        {
+            var Index = CurrentApiUiService!.OrderList!.FindIndex(Order =>
+            Order.PropertyName != null
+            && Order.PropertyName.Equals(PropertyName, StringComparison.OrdinalIgnoreCase));
+            if (Index > -1)
+            {
+                SortOrderDirection OrderType = SortOrderDirection.None;
+                switch (OrderTypeValue.ToLower())
+                {
+                    case "ascending":
+                        OrderType = SortOrderDirection.Ascending;
+                        break;
+                    case "descending":
+                        OrderType = SortOrderDirection.Descending;
+                        break;
+                }
+                CurrentApiUiService?.OrderList?[Convert.ToInt32(Index)].UpdateCurrentOrderDirection(OrderType);
+            }
+
+            CurrentApiUiService!.SwitchOffOtherOrderList(PropertyName);
+
+            if ({{entity.Name}}DataGridTable != null)
+            {
+#pragma warning disable BL0005 // Component parameter should not be set outside of its component. 
+                {{entity.Name}}DataGridTable!.CurrentPage = 0; //only reliable way to reset page on reorder
+#pragma warning restore BL0005 // Component parameter should not be set outside of its component.
+                await {{entity.Name}}DataGridTable!.ReloadServerData();
+            }
+        }
+
+        IsLoading = false;
+    }
+
+    public SortDirection GetPropertyMudSortDirection(string PropertyName)
+    {
+        SortDirection GetSortDirection = SortDirection.None;
+
+        if (IsOrderPopulated)
+        {
+            switch (CurrentApiUiService!.GetCurrentOrderType(PropertyName))
+            {
+                case SortOrderDirection.Ascending:
+                    GetSortDirection = SortDirection.Ascending;
+                    break;
+                case SortOrderDirection.Descending:
+                    GetSortDirection = SortDirection.Descending;
+                    break;
+            }
+        }
+
+        return GetSortDirection;
+    }
+
+    #endregion
+
+    #region Delete
+
     public async Task DeleteOnClick({{entity.Name}}Model? currentSelection)
     {
         if (currentSelection != null)
@@ -54,4 +209,6 @@ public partial class {{className}} : ComponentBase
             await OnDeleteChanged.InvokeAsync(currentSelection);
         }        
     }
+
+    #endregion
 }
